+++ SignNow.Net/Model/Responses/UserCreateResponse.cs
@@ -1,0 +1,28 @@
+using Newtonsoft.Json;
+
+namespace SignNow.Net.Model
+{
+    /// <summary>
+    /// Represents response from SignNow API for User create request.
+    /// </summary>
+    public class UserCreateResponse
+    {
+        /// <summary>
+        /// User identity.
+        /// </summary>
+        [JsonProperty("id")]
+        public string Id { get; set; }
+
+        /// <summary>
+        /// User is verified or not.
+        /// </summary>
+        [JsonProperty("verified")]
+        public bool Verified { get; set; }
+
+        /// <summary>
+        /// User email.
+        /// </summary>
+        [JsonProperty("email")]
+        public string Email { get; set; }
+    }
+}