+++ SignNow.Net/Service/UserService.cs
@@ -8,6 +8,7 @@ using SignNow.Net.Internal.Extensions;
 using SignNow.Net.Internal.Helpers;
 using SignNow.Net.Internal.Requests;
 using SignNow.Net.Model;
+using SignNow.Net.Model.Requests;
 
 namespace SignNow.Net.Service
 {
@@ -25,6 +26,23 @@ namespace SignNow.Net.Service
         {
         }
 
+        /// <inheritdoc cref="IUserService.CreateUserAsync" />
+        public async Task<UserCreateResponse> CreateUserAsync(CreateUserOptions createUser, CancellationToken cancellation = default)
+        {
+            var basicToken = Token;
+            basicToken.TokenType = TokenType.Basic;
+
+            var requestOptions = new PostHttpRequestOptions
+            {
+                RequestUrl = new Uri(ApiBaseUrl, "/user"),
+                Content = new JsonHttpContent(createUser),
+                Token = basicToken
+            };
+
+            return await SignNowClient.RequestAsync<UserCreateResponse>(requestOptions, cancellation)
+                .ConfigureAwait(false);
+        }
+
         /// <inheritdoc cref="IUserService.GetCurrentUserAsync" />
         public async Task<User> GetCurrentUserAsync(CancellationToken cancellationToken = default)
         {
@@ -37,8 +55,64 @@ namespace SignNow.Net.Service
             return await SignNowClient.RequestAsync<User>(requestOptions, cancellationToken).ConfigureAwait(false);
         }
 
+        /// <inheritdoc cref="IUserService.UpdateUserAsync" />
+        public async Task<UserUpdateResponse> UpdateUserAsync(UpdateUserOptions updateUser, CancellationToken cancellationToken = default)
+        {
+            var requestOptions = new PutHttpRequestOptions
+            {
+                RequestUrl = new Uri(ApiBaseUrl, "/user"),
+                Content = new JsonHttpContent(updateUser),
+                Token = Token
+            };
+
+            return await SignNowClient
+                .RequestAsync<UserUpdateResponse>(requestOptions, cancellationToken)
+                .ConfigureAwait(false);
+        }
+
+        /// <inheritdoc cref="IUserService.SendVerificationEmailAsync" />
+        /// <exception cref="ArgumentException"><paramref name="email"/> address is not valid</exception>
+        public async Task SendVerificationEmailAsync(string email, CancellationToken cancellationToken = default)
+        {
+            var basicToken = Token;
+            basicToken.TokenType = TokenType.Basic;
+
+            var requestBody = new {email = email.ValidateEmail()};
+
+            var requestOptions = new PostHttpRequestOptions
+            {
+                RequestUrl = new Uri(ApiBaseUrl, "/user/verifyemail"),
+                Content = new JsonHttpContent(requestBody),
+                Token = basicToken
+            };
+
+            await SignNowClient.RequestAsync(requestOptions, cancellationToken)
+                .ConfigureAwait(false);
+        }
+
+        /// <inheritdoc cref="IUserService.SendPasswordResetLinkAsync"/>
+        /// <exception cref="ArgumentException"><paramref name="email"/> address is not valid</exception>
+        public async Task SendPasswordResetLinkAsync(string email, CancellationToken cancellationToken = default)
+        {
+            var basicToken = Token;
+            basicToken.TokenType = TokenType.Basic;
+
+            var requestBody = new {email = email.ValidateEmail()};
+
+            var requestOptions = new PostHttpRequestOptions
+            {
+                RequestUrl = new Uri(ApiBaseUrl, "/user/forgotpassword"),
+                Content = new JsonHttpContent(requestBody),
+                Token = basicToken
+            };
+
+            await SignNowClient.RequestAsync(requestOptions, cancellationToken)
+                .ConfigureAwait(false);
+        }
+
         /// <inheritdoc cref="ISignInvite.CreateInviteAsync" />
-        /// <exception cref="ArgumentNullException"><see cref="SignInvite"/> cannot be null.</exception>
+        /// <exception cref="ArgumentNullException"><paramref name="invite"/> cannot be null.</exception>
+        /// <exception cref="ArgumentException">Invalid format of <paramref name="documentId"/></exception>
         public async Task<InviteResponse> CreateInviteAsync(string documentId, SignInvite invite, CancellationToken cancellationToken = default)
         {
             Guard.ArgumentNotNull(invite, nameof(invite));
@@ -59,7 +133,7 @@ namespace SignNow.Net.Service
         }
 
         /// <inheritdoc cref="ISignInvite.CancelInviteAsync(FreeformInvite, CancellationToken)" />
-        /// <exception cref="ArgumentNullException"><see cref="FreeformInvite"/> cannot be null.</exception>
+        /// <exception cref="ArgumentNullException"><paramref name="invite"/> cannot be null.</exception>
         public async Task CancelInviteAsync(FreeformInvite invite, CancellationToken cancellationToken = default)
         {
             Guard.ArgumentNotNull(invite, nameof(invite));
@@ -68,7 +142,7 @@ namespace SignNow.Net.Service
         }
 
         /// <inheritdoc cref="ISignInvite.CancelInviteAsync(string, CancellationToken)" />
-        /// <exception cref="ArgumentException">Invalid format of Document Id.</exception>
+        /// <exception cref="ArgumentException">Invalid format of <paramref name="documentId"/>.</exception>
         public async Task CancelInviteAsync(string documentId, CancellationToken cancellationToken = default)
         {
             await ProcessCancelInviteAsync($"/document/{documentId.ValidateId()}/fieldinvitecancel", cancellationToken).ConfigureAwait(false);